requirements.txt
Django==4.1.5
environs==7.3.1
django_compressor==4.3
django-libsass==0.9
django-filter==22.1
